[KAN-62] refactor: /gathering 페이지 컴포넌트 분리
